+++ CHANGELOG.md
@@ -11,6 +11,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - response body is now always readable even when providing a buffer or struct
 
+### 1.5.0 - 01-06-2023
+
+### Added
+
+- add trailing slash to base url if not given
+
+### 1.4.0 - 05-03-2023
+
+- upgrade dependencies
+
+### 1.3.0 - 09-12-2020
+
+#### Added
+
+- Unmarshal method as `HTTPError` method
+- expose Raw as slice of bytes in `HTTPError`
+- set an Host header to all client requests
+
+### 1.2.0 - 11-10-2020
+
+#### Added
+
+- add custom http client to perform http request
+
 ### 1.1.0 - 01-06-2020
 
 #### Added